Rename ta-librett.h to librett.h
